# backend/datahub/openapi/datahub-v1.yaml
openapi: 3.1.0
info:
  title: AutoJMS DataHub API
  version: 1.0.0
  description: |
    Phase-1 API contract for the VPS-hosted DataHub. PostgreSQL is private to the
    API deployment and clients use REST plus SignalR. The signed DataHub license
    channel is independent from the desktop update channel.

    `scanTime` is the only business timestamp. A naive JMS value in
    `yyyy-MM-dd HH:mm:ss` is interpreted as `Asia/Ho_Chi_Minh`; an ISO value with
    `Z` or an explicit offset is interpreted using that offset. `uploadTime` is
    accepted only inside the raw `payload` object and is never used for ordering,
    fingerprinting, projection reduction, or retention.
  license:
    name: Internal use
    identifier: internal
servers:
  - url: https://datahub.example.com
    description: Production DataHub (the DNS target can move without changing a license)
  - url: https://datahub-dev.example.com
    description: Staging DataHub (independent database and credentials)
...
paths:
...
  /api/v1/sites/{siteId}/lease/acquire:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    post:
      tags: [Lease]
      operationId: acquireSiteLease
      summary: Acquire an expired site bulk-fetch lease
      description: |
        The authenticated device is the only device identity considered. The path
        `siteId` must match the device token claim. A request body, body `siteId`,
        body/header channel, or caller-supplied device ID cannot grant access.
        Acquire/steal increments the monotonic `leaderTerm`; renewing the current
        term does not. A lease row is never deleted.
      security:
        - DeviceBearer: []
      x-authorization-source: device-token-claims
      responses:
        '200':
          description: The caller is now the leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Channel or site claim does not match the deployment/path
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                channelMismatch:
                  $ref: '#/components/examples/ChannelMismatch'
                siteNotLicensed:
                  $ref: '#/components/examples/SiteNotLicensed'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/LeaseHeld'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sites/{siteId}/lease/renew:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    post:
      tags: [Lease]
      operationId: renewSiteLease
      summary: Renew the caller's current lease term
      description: |
        The request must carry the term currently held by the authenticated device.
        Renew does not increment the term. An expired or superseded term is fenced.
      security:
        - DeviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseTermRequest'
      responses:
        '200':
          description: Lease renewed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/LeaderFenced'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sites/{siteId}/lease/release:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    post:
      tags: [Lease]
      operationId: releaseSiteLease
      summary: Release the caller's lease and fence in-flight bulk requests
      description: |
        Release increments `leaderTerm`, clears `leaderDeviceId`, and sets the
        expiry to `-infinity`. It never deletes the lease row. Requests carrying the
        old term receive `LEADER_FENCED`.
      security:
        - DeviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeaseTermRequest'
      responses:
        '200':
          description: Lease released and old term fenced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaseState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/LeaderFenced'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/v1/sites/{siteId}/changes:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    get:
      tags: [Sync]
      operationId: readDashboardChanges
      summary: Read committed per-site changes after a client cursor
      description: |
        Returns `changeSeq > after`, ordered by `changeSeq`. The API never requires
        `after + 1`, and SignalR is not the data transport. Each upsert body is the
        complete hot projection snapshot needed for one SQLite apply transaction.
        If `after` is older than the retained change range, return `RESYNC_REQUIRED`
        and the client must call the snapshot endpoint.
      security:
        - DeviceBearer: []
      parameters:
        - $ref: '#/components/parameters/After'
        - $ref: '#/components/parameters/ChangeLimit'
      responses:
        '200':
          description: Ordered delta page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResyncRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/sites/{siteId}/projections/snapshot:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    get:
      tags: [Sync]
      operationId: readProjectionSnapshot
      summary: Read a consistent site projection snapshot
      description: |
        Phase 1 returns the complete snapshot as one HTTP response. The server
        opens one PostgreSQL `REPEATABLE READ` transaction, reads one
        `snapshot_seq` watermark from `site_change_counters`, reads all projection
        rows in that transaction, and commits before writing the response. All
        items in this response therefore share the returned `snapshot_seq`. Snapshot
        tokens and multi-request paging are intentionally deferred to a later phase.
        The client applies the response, sets its cursor to `snapshot_seq`, then calls
        `/changes?after=snapshot_seq`.
      security:
        - DeviceBearer: []
      x-transaction-isolation: REPEATABLE READ
      x-snapshot-watermark: snapshot_seq
      responses:
        '200':
          description: Complete consistent projection snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    LicenseAssertion:
      type: http
      scheme: bearer
      bearerFormat: signed-license-assertion
      description: |
        Signed DataHub license assertion accepted only by `POST /api/v1/devices/enroll`.
        It carries the required `channel` (`staging` or `production`), `site_codes`,
        expiry, and optionally a signed HTTPS `datahub_url`. It never carries JMS
        tokens, PostgreSQL credentials, DPAPI blobs, or device secrets. Production
        uses asymmetric JWS verification. The custom HMAC assertion is an explicitly
        enabled staging integration seam only; production readiness remains unhealthy
        until the asymmetric verifier is installed.
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: device-token
      description: |
        Derived device credential returned by enrollment. Claims include `device_id`,
        `site_id`, `channel`, role, and `token_version`. It is required for every
        endpoint except health and enrollment. The API rejects a channel claim that
        differs from `DATAHUB_CHANNEL` and never trusts a raw channel/site header or
        body field.
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      description: Immutable site UUID. Must equal the authenticated device claim.
      schema:
        type: string
        format: uuid
...
    After:
      name: after
      in: query
      required: false
      description: Return changes with `changeSeq > after` for this site.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
    ChangeLimit:
      name: limit
      in: query
      required: false
      description: Maximum number of changes in one bounded response.
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 500
  schemas:
...
    LeaseTermRequest:
      type: object
      additionalProperties: false
      required: [leaderTerm]
      properties:
        leaderTerm:
          type: integer
          format: int64
          minimum: 1
    LeaseState:
      type: object
      required: [siteId, leaderDeviceId, leaderTerm, leaseExpiresAt, role]
      properties:
        siteId:
          type: string
          format: uuid
        leaderDeviceId:
          type: [string, 'null']
          format: uuid
        leaderTerm:
          type: integer
          format: int64
          minimum: 0
        leaseExpiresAt:
          type: [string, 'null']
          format: date-time
          description: "The -infinity database value is represented by the API as null before first acquisition or after release."
        lastSeenAt:
          type: [string, 'null']
          format: date-time
        role:
          type: string
          enum: [leader, follower, released]
        leaseDurationSeconds:
          type: integer
          const: 120
        renewIntervalSeconds:
          type: integer
          const: 30
...
    ProjectionSnapshot:
      type: object
      required: [siteId, waybillNo, payload, reducerVersion, version, updatedAt]
      properties:
        siteId:
          type: string
          format: uuid
        waybillNo:
          type: string
        stateCode:
          type: [integer, 'null']
        stateName:
          type: [string, 'null']
        stateStatus:
          type: [string, 'null']
        stateEventAt:
          type: [string, 'null']
          format: date-time
        stateFingerprint:
          type: [string, 'null']
        stateEventId:
          type: [integer, 'null']
          format: int64
        stateKind:
          type: [string, 'null']
        lastActivityCode:
          type: [integer, 'null']
        lastActivityName:
          type: [string, 'null']
        lastActivityStatus:
          type: [string, 'null']
        lastActivityKind:
          type: [string, 'null']
        lastActivityAt:
          type: [string, 'null']
          format: date-time
        lastActivityFingerprint:
          type: [string, 'null']
        lastActivityEventId:
          type: [integer, 'null']
          format: int64
        inventoryCode:
          type: [integer, 'null']
        inventoryName:
          type: [string, 'null']
        inventoryKind:
          type: [string, 'null']
          const: inventory
        inventoryStatus:
          type: [string, 'null']
        inventoryEventAt:
          type: [string, 'null']
          format: date-time
        inventoryFingerprint:
          type: [string, 'null']
        inventoryEventId:
          type: [integer, 'null']
          format: int64
        payload:
          type: object
          additionalProperties: true
          default: {}
          description: Compact, apply-safe hot snapshot; not the full JMS envelope.
        statePayload:
          type: [object, 'null']
          additionalProperties: true
          description: Retained compact payload for the winning state event.
        activityPayload:
          type: [object, 'null']
          additionalProperties: true
          description: Retained compact payload for the winning activity event.
        inventoryPayload:
          type: [object, 'null']
          additionalProperties: true
          description: Retained compact payload for the winning inventory event.
        reducerVersion:
          type: integer
          minimum: 1
        version:
          type: integer
          format: int64
          minimum: 1
        updatedAt:
          type: string
          format: date-time
    Tombstone:
      type: object
      required: [deleted, siteId, waybillNo, version]
      properties:
        deleted:
          type: boolean
          const: true
        siteId:
          type: string
          format: uuid
        waybillNo:
          type: string
        version:
          type: integer
          format: int64
    DashboardChange:
      type: object
      required: [siteId, changeSeq, entityType, entityKey, operation, changeAt, body]
      properties:
        siteId:
          type: string
          format: uuid
        changeSeq:
          type: integer
          format: int64
          minimum: 1
        entityType:
          type: string
          const: waybill_projection
        entityKey:
          type: string
        operation:
          type: string
          enum: [upsert, delete, resync]
        changeAt:
          type: string
          format: date-time
          description: Server synchronization timestamp; never the business event timestamp.
        body:
          oneOf:
            - $ref: '#/components/schemas/ProjectionSnapshot'
            - $ref: '#/components/schemas/Tombstone'
          description: Complete hot projection body for upsert, or an applicable tombstone.
    ChangePage:
      type: object
      required: [siteId, after, items, hasMore, nextAfter]
      properties:
        siteId:
          type: string
          format: uuid
        after:
          type: integer
          format: int64
          minimum: 0
        items:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/DashboardChange'
        hasMore:
          type: boolean
        nextAfter:
          type: integer
          format: int64
          minimum: 0
    SnapshotResponse:
      type: object
      required: [siteId, snapshot_seq, items, itemCount]
      properties:
        siteId:
          type: string
          format: uuid
        snapshot_seq:
          type: integer
          format: int64
          minimum: 0
          description: Per-site change watermark captured at the start of the REPEATABLE READ transaction.
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectionSnapshot'
        itemCount:
          type: integer
          minimum: 0
        generatedAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [type, title, status, code, detail, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          type: string
        traceId:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - INVALID_SCAN_TIME
        - UNAUTHORIZED
        - FORBIDDEN
        - CHANNEL_MISMATCH
        - SITE_NOT_LICENSED
        - NOT_FOUND
        - LEADER_FENCED
        - LEASE_HELD
        - SEAT_LIMIT_REACHED
        - DEVICE_CONFLICT
        - IDEMPOTENCY_KEY_REUSED
        - IDEMPOTENCY_IN_PROGRESS
        - RESYNC_REQUIRED
        - PAYLOAD_TOO_LARGE
        - VALIDATION_FAILED
        - RATE_LIMITED
        - SERVICE_UNAVAILABLE
  examples:
    ChannelMismatch:
      summary: License/device channel differs from DATAHUB_CHANNEL
      value:
        type: https://datahub.example.com/problems/channel-mismatch
        title: DataHub channel mismatch
        status: 403
        code: CHANNEL_MISMATCH
        detail: The signed license or device token channel is not accepted by this deployment.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
    SiteNotLicensed:
      summary: Site is outside signed site_codes
      value:
        type: https://datahub.example.com/problems/site-not-licensed
        title: Site is not licensed
        status: 403
        code: SITE_NOT_LICENSED
        detail: The requested site is not present in the signed license site_codes claim.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
...
    LeaderFenced:
      summary: Stale or invalid leader term
      value:
        type: https://datahub.example.com/problems/leader-fenced
        title: Leader fenced
        status: 409
        code: LEADER_FENCED
        detail: The device, term, or lease expiry no longer matches the current leader.
        traceId: 00-00000000000000000000000000000000-0000000000000000-00
...
  responses:
...
    Unauthorized:
      description: Missing, expired, malformed, or invalid bearer credential
      headers:
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated but not authorized for the deployment or site
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Site, device, or requested resource was not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    LeaseHeld:
      description: Another device currently holds the unexpired site lease
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            leaseHeld:
              value:
                type: https://datahub.example.com/problems/lease-held
                title: Lease held
                status: 409
                code: LEASE_HELD
                detail: Another enrolled device holds the site bulk-fetch lease.
                traceId: 00-00000000000000000000000000000000-0000000000000000-00
    LeaderFenced:
      description: The supplied lease term is stale or no longer owned by this device
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            leaderFenced:
              $ref: '#/components/examples/LeaderFenced'
...
    ResyncRequired:
      description: Cursor is older than retained changes; take a consistent snapshot
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Caller exceeded the endpoint rate limit
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: API or required PostgreSQL dependency is unavailable
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'